requirements.txt
requests>=2.31.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
flask>=3.0.0
supabase>=2.0.0
